[cresp_crspectrum,cresp_grid] FIX: added precaution for second solution failure + counter, bad interpolation precaution.
In case when no solution is found after second try p_?? = p_mid_fix is
assumed. Count of events seems to be high compared to cases solution failures
in the first try. In "get_fqp_.." added condition that assures ratio > 1,
other results are due to bad interpolation (swap zeros with ones on the
solution map?).